Refactor checkpointing
